Resolve compile warnings in modern

+++ include/task.h
@@ -33,7 +33,7 @@ void SwitchTaskToFollowupFunc(u8 taskId);
 bool8 FuncIsActiveTask(TaskFunc func);
 u8 FindTaskIdByFunc(TaskFunc func);
 u8 GetTaskCount(void);
-void SetWordTaskArg(u8 taskId, u8 dataElem, u32 value);
+void SetWordTaskArg(u8 taskId, u8 dataElem, unsigned long value);
 u32 GetWordTaskArg(u8 taskId, u8 dataElem);
 
 #endif // GUARD_TASK_H
